created msleep

+++ philo/srcs/now.c
@@ -6,13 +6,13 @@
 /*   By: pcunha <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/30 17:11:11 by pcunha            #+#    #+#             */
-/*   Updated: 2021/07/30 18:15:47 by pcunha           ###   ########.fr       */
+/*   Updated: 2021/07/30 18:19:09 by pcunha           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	now(void)
+long long int	now(void)
 {
 	t_timeval time;
